+++ src/pages/payments/register/flow/Flow.tsx
@@ -1,0 +1,139 @@
+import { Outlet } from 'react-router';
+
+import { useCardNumbers } from '../form/hooks/useCardNumbers';
+import { useCard } from '../form/hooks/useCard';
+import { useExpirationDate } from '../form/hooks/useExpirationDate';
+import { useCvc } from '../form/hooks/useCvc';
+import { usePassword } from '../form/hooks/usePassword';
+
+import { BRAND_NUMBER } from '../form/constant';
+
+export const Flow = () => {
+  const {
+    value: cardNumbers,
+    onChange: handleChangeCardNumbers,
+
+    onBlur: handleBlurCardNumbers,
+
+    renderErrorMessage: renderErrorMessageCardNumbers,
+    renderErrorInput: renderErrorCardNumberInput,
+  } = useCardNumbers();
+
+  const {
+    value: card,
+    onChange: handleChangeCard,
+
+    onBlur: handleBlurCard,
+
+    renderErrorMessage: renderErrorMessageCard,
+  } = useCard();
+
+  const {
+    value: expirationDate,
+    onChange: handleChangeExpirationDate,
+
+    blurValue: onBlurExpirationDate,
+    onBlur: handleBlurExpirationDate,
+
+    errors: errorsExpirationDate,
+
+    invalidAttemp: expirationDateInvalidAttemp,
+    renderErrorMessage: renderErrorMessageExpirationDate,
+  } = useExpirationDate();
+
+  //cvc
+  const {
+    value: cvc,
+    onChange: handleChangeCvc,
+
+    blurValue: onBlurCvc,
+    onBlur: handleBlurCvc,
+
+    errors: errorsCvc,
+
+    invalidAttemp: cvcInvalidAttemp,
+    renderErrorMessage: renderErrorMessageCvc,
+  } = useCvc();
+
+  //password
+  const {
+    value: password,
+    onChange: handleChangePassword,
+
+    blurValue: onBlurPassword,
+    onBlur: handleBlurPassword,
+
+    errors: errorsPassword,
+
+    invalidAttemp: passwordInvalidAttemp,
+    renderErrorMessage: renderErrorMessagePassword,
+  } = usePassword();
+
+  const renderBrandCard = (cardNumbers: string[]) => {
+    if (cardNumbers[0].startsWith(BRAND_NUMBER.visa)) return 'visa';
+    if (BRAND_NUMBER.mastercard.some((brandNumber) => cardNumbers[0].startsWith(brandNumber))) return 'mastercard';
+    return 'default';
+  };
+
+  return (
+    <Outlet
+      context={{
+        // cardNumbers
+        cardNumbers,
+        handleChangeCardNumbers,
+
+        handleBlurCardNumbers,
+
+        renderErrorMessageCardNumbers,
+        renderErrorCardNumberInput,
+
+        // card
+        card,
+        handleChangeCard,
+
+        handleBlurCard,
+
+        renderErrorMessageCard,
+
+        // expirationDate
+        expirationDate,
+        handleChangeExpirationDate,
+
+        onBlurExpirationDate,
+        handleBlurExpirationDate,
+
+        errorsExpirationDate,
+
+        expirationDateInvalidAttemp,
+        renderErrorMessageExpirationDate,
+
+        // cvc
+        cvc,
+        handleChangeCvc,
+
+        onBlurCvc,
+        handleBlurCvc,
+
+        errorsCvc,
+
+        cvcInvalidAttemp,
+        renderErrorMessageCvc,
+
+        // password
+        password,
+        handleChangePassword,
+
+        onBlurPassword,
+        handleBlurPassword,
+
+        errorsPassword,
+
+        passwordInvalidAttemp,
+        renderErrorMessagePassword,
+
+        // creditCard
+        renderBrandCard,
+      }}
+    />
+  );
+};
