+++ src/pages/payments/register/flow/index.ts
@@ -1,0 +1,1 @@
+export { Flow } from './Flow';
