+++ src/routes/AppRoutes.tsx
@@ -1,14 +1,16 @@
 import { Routes, Route } from 'react-router';
 
+import { Flow } from '@/pages/payments/register/flow';
 import { Form } from '@/pages/payments/register/form';
 import { Complete } from '@/pages/payments/register/complete';
 
 export const AppRoutes = () => {
   return (
     <Routes>
-      <Route path="/" element={<Form />} />
-      <Route path="/payments/register" element={<Form />} />
-      <Route path="/payments/complete" element={<Complete cardNumber="5511" card="bc" />} />
+      <Route path="/payments/register" element={<Flow />}>
+        <Route index element={<Form />} />
+        <Route path="complete" element={<Complete cardNumber="5511" card="bc" />} />
+      </Route>
     </Routes>
   );
 };
